game logic improving and screen result design

+++ app/src/main/java/newagency/picfav/view/main/presenter/model/GameResult.java
@@ -6,6 +6,6 @@ package newagency.picfav.view.main.presenter.model;
 
 public class GameResult {
 
-    public float score;
+    public int score;
 
 }
